Split up actor methods Split up bookmark methods Move actions into helper folders in the appropriate namespaces
