+++ CHANGELOG.md
@@ -10,6 +10,7 @@
   for enumeration of an authenticated user's opt-outs, and `DeleteReferenceListOptOut` allows the deletion of the opt-out. Bluesky allows multiple optout
   records for the same list. If another opt-out record for the list still exists after deletion the opt-out remains in effect. See [Add reference list opt-out lexicons](https://github.com/bluesky-social/indigo/pull/1447/changes).
 * `ListBlock` has been added in the `idunno.Bluesky.Graph` namespace, representing a block relationship against an entire an entire list of accounts (actors).
+* Added `Actor.ProfileAssociatedActivitySubscription` and the known values `AllowSubscriptionsKnownValues`. This appears in the lexicon, but does not current seem to be used anywhere.
 
 ### Breaking Changes
