+++ src/idunno.Bluesky/Actor/Agent/Helpers/PutPreference.cs
@@ -1,0 +1,30 @@
+// Copyright (c) Barry Dorrans. All rights reserved.
+// Licensed under the MIT License.
+
+using idunno.AtProto;
+using idunno.Bluesky.Actor;
+
+namespace idunno.Bluesky;
+
+public partial class BlueskyAgent
+{
+    /// <summary>
+    /// Updates the specified preference for the current user.
+    /// </summary>
+    /// <param name="preference">The preference to update</param>
+    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>The task object representing the asynchronous operation.</returns>
+    /// <exception cref="ArgumentNullException">Thrown when <paramref name="preference"/> is <see langword="null"/>.</exception>
+    /// <exception cref="AuthenticationRequiredException">Thrown when the current agent is not authenticated.</exception>
+    public async Task<AtProtoHttpResult<EmptyResponse>> PutPreference(Preference preference, CancellationToken cancellationToken = default)
+    {
+        ArgumentNullException.ThrowIfNull(preference);
+
+        if (!IsAuthenticated)
+        {
+            throw new AuthenticationRequiredException();
+        }
+
+        return await PutPreferences(preference, cancellationToken: cancellationToken).ConfigureAwait(false);
+    }
+}
