+++ src/idunno.Bluesky/Actor/ContentVisibilityDeclaration.cs
