+++ src/idunno.Bluesky/Actor/Defs/ProfileViewDetailed.cs
