+++ src/idunno.Bluesky/Actor/Server/GetProfile.cs
@@ -1,0 +1,64 @@
+// Copyright (c) Barry Dorrans. All rights reserved.
+// Licensed under the MIT License.
+
+using System.Diagnostics.CodeAnalysis;
+
+using idunno.AtProto;
+using idunno.AtProto.Authentication;
+using idunno.Bluesky.Actor;
+
+using Microsoft.Extensions.Logging;
+
+namespace idunno.Bluesky;
+
+public static partial class BlueskyServer
+{
+    /// <summary>
+    /// Get a detailed profile view of an account. Does not require authentication,
+    /// but if an access token is provided will contain relevant metadata about the requesting
+    /// actor's relationship to the requested account.
+    /// </summary>
+    /// <param name="actor">The handle or DID of the account to fetch profile of.</param>
+    /// <param name="service">The <see cref="Uri"/> of the service to retrieve the profile from.</param>
+    /// <param name="accessCredentials">The <see cref="AccessCredentials"/> used to authenticate to <paramref name="service"/>.</param>
+    /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
+    /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the credentials in the request need updating.</param>
+    /// <param name="loggerFactory">An instance of <see cref="ILoggerFactory"/> to use to create a logger.</param>
+    /// <param name="subscribedLabelers">An optional list of <see cref="Did"/>s of labelers to retrieve labels applied to the account.</param>
+    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>The task object representing the asynchronous operation.</returns>
+    /// <exception cref="ArgumentNullException">Thrown when any of <paramref name="actor"/>, <paramref name="service"/> or <paramref name="httpClient"/> are <see langword="null"/>.</exception>
+    [UnconditionalSuppressMessage(
+        "Trimming",
+        "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
+        Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
+    [UnconditionalSuppressMessage("AOT",
+        "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
+        Justification = "All types are preserved in the JsonSerializerOptions call to Get().")]
+    public static async Task<AtProtoHttpResult<ProfileViewDetailed>> GetProfile(
+        AtIdentifier actor,
+        Uri service,
+        AccessCredentials? accessCredentials,
+        HttpClient httpClient,
+        Action<AtProtoCredential>? onCredentialsUpdated = null,
+        ILoggerFactory? loggerFactory = default,
+        IEnumerable<Did>? subscribedLabelers = null,
+        CancellationToken cancellationToken = default)
+    {
+        ArgumentNullException.ThrowIfNull(actor);
+        ArgumentNullException.ThrowIfNull(service);
+        ArgumentNullException.ThrowIfNull(httpClient);
+
+        BlueskyHttpClient<ProfileViewDetailed> request = new(AppViewProxy, loggerFactory);
+
+        return await request.Get(
+            service,
+            $"/xrpc/app.bsky.actor.getProfile?actor={Uri.EscapeDataString(actor.ToString())}",
+            credentials: accessCredentials,
+            onCredentialsUpdated: onCredentialsUpdated,
+            httpClient: httpClient,
+            jsonSerializerOptions: BlueskyJsonSerializerOptions,
+            subscribedLabelers: subscribedLabelers,
+            cancellationToken: cancellationToken).ConfigureAwait(false);
+    }
+}
