+++ src/idunno.Bluesky/Bookmarks/Server/CreateBookmark.cs
@@ -1,0 +1,74 @@
+// Copyright (c) Barry Dorrans. All rights reserved.
+// Licensed under the MIT License.
+
+using System.Diagnostics.CodeAnalysis;
+
+using idunno.AtProto;
+using idunno.AtProto.Authentication;
+using idunno.Bluesky.Bookmarks.Model;
+
+using Microsoft.Extensions.Logging;
+
+namespace idunno.Bluesky;
+
+public static partial class BlueskyServer
+{
+    // Bookmarks - https://github.com/bluesky-social/atproto/pull/4163
+
+    /// <summary>
+    /// Creates a bookmark on the specified account from the specified <paramref name="uri"/> and <paramref name="cid"/>.
+    /// </summary>
+    /// <param name="uri">The <see cref="AtUri"/> of the post to bookmark</param>
+    /// <param name="cid">The <see cref="Cid"/> of the post to bookmark</param>
+    /// <param name="service">The <see cref="Uri"/> of the service to add the bookmark to.</param>
+    /// <param name="accessCredentials">The <see cref="AccessCredentials"/> used to authenticate to <paramref name="service"/>.</param>
+    /// <param name="httpClient">An <see cref="HttpClient"/> to use when making a request to the <paramref name="service"/>.</param>
+    /// <param name="onCredentialsUpdated">An <see cref="Action{T}" /> to call if the credentials in the request need updating.</param>
+    /// <param name="loggerFactory">An instance of <see cref="ILoggerFactory"/> to use to create a logger.</param>
+    /// <param name="cancellationToken">A cancellation token that can be used by other objects or threads to receive notice of cancellation.</param>
+    /// <returns>The task object representing the asynchronous operation.</returns>
+    /// <exception cref="ArgumentNullException">Thrown when any of <paramref name="uri"/>, <paramref name="cid"/>, <paramref name="service"/>,<paramref name="accessCredentials"/> or <paramref name="httpClient"/> are <see langword="null"/>.</exception>
+    /// <exception cref="ArgumentOutOfRangeException">Thrown when <paramref name="uri"/> does not point to a post</exception>
+    [UnconditionalSuppressMessage(
+        "Trimming",
+        "IL2026:Members annotated with 'RequiresUnreferencedCodeAttribute' require dynamic access otherwise can break functionality when trimming application code",
+        Justification = "All types are preserved in the JsonSerializerOptions call to Post().")]
+    [UnconditionalSuppressMessage("AOT",
+        "IL3050:Calling members annotated with 'RequiresDynamicCodeAttribute' may break functionality when AOT compiling.",
+        Justification = "All types are preserved in the JsonSerializerOptions call to Post().")]
+    public static async Task<AtProtoHttpResult<EmptyResponse>> CreateBookmark(
+        AtUri uri,
+        Cid cid,
+        Uri service,
+        AccessCredentials accessCredentials,
+        HttpClient httpClient,
+        Action<AtProtoCredential>? onCredentialsUpdated = null,
+        ILoggerFactory? loggerFactory = default,
+        CancellationToken cancellationToken = default)
+    {
+        ArgumentNullException.ThrowIfNull(uri);
+        ArgumentNullException.ThrowIfNull(cid);
+        ArgumentNullException.ThrowIfNull(uri.Collection);
+        ArgumentOutOfRangeException.ThrowIfNotEqual(uri.Collection, CollectionNsid.Post);
+
+        ArgumentNullException.ThrowIfNull(service);
+        ArgumentNullException.ThrowIfNull(accessCredentials);
+        ArgumentNullException.ThrowIfNull(httpClient);
+
+        BlueskyHttpClient<EmptyResponse> client = new(AppViewProxy, loggerFactory);
+
+        CreateBookmarkRequest request = new(uri, cid);
+
+        AtProtoHttpResult<EmptyResponse> response = await client.Post(
+            service,
+            "/xrpc/app.bsky.bookmark.createBookmark",
+            request,
+            credentials: accessCredentials,
+            httpClient: httpClient,
+            jsonSerializerOptions: BlueskyJsonSerializerOptions,
+            onCredentialsUpdated: onCredentialsUpdated,
+            cancellationToken: cancellationToken).ConfigureAwait(false);
+
+        return response;
+    }
+}
